[25/n][eas-cli] Move graphql and session into context
