+++ packages/eas-cli/src/api.ts
@@ -1,93 +1,78 @@
 import { JSONValue } from '@expo/json-file';
 
+import { ApiV2Error } from './ApiV2Error';
 import fetch, { RequestError, RequestInit } from './fetch';
-import { getAccessToken, getSessionSecret } from './user/sessionStorage';
-
-export class ApiV2Error extends Error {
-  override readonly name = 'ApiV2Error';
-  readonly expoApiV2ErrorCode: string;
-  readonly expoApiV2ErrorDetails?: JSONValue;
-  readonly expoApiV2ErrorServerStack?: string;
-  readonly expoApiV2ErrorMetadata?: object;
-
-  constructor(response: {
-    message: string;
-    code: string;
-    stack?: string;
-    details?: JSONValue;
-    metadata?: object;
-  }) {
-    super(response.message);
-    this.expoApiV2ErrorCode = response.code;
-    this.expoApiV2ErrorDetails = response.details;
-    this.expoApiV2ErrorServerStack = response.stack;
-    this.expoApiV2ErrorMetadata = response.metadata;
-  }
-}
 
 interface RequestOptions {
   body: JSONValue;
 }
 
-export const api = {
-  async putAsync(path: string, options: RequestOptions): Promise<any> {
-    return await requestAsync(path, { method: 'PUT', body: JSON.stringify(options.body) });
-  },
+export class ApiV2Client {
+  constructor(
+    private readonly authInfo: {
+      accessToken: string | null;
+      sessionSecret: string | null;
+    }
+  ) {}
 
-  async postAsync(path: string, options: RequestOptions): Promise<any> {
-    return await requestAsync(path, { method: 'POST', body: JSON.stringify(options.body) });
-  },
-
-  async deleteAsync(path: string): Promise<any> {
-    return await requestAsync(path, { method: 'DELETE' });
-  },
-
-  async getAsync(path: string): Promise<any> {
-    return await requestAsync(path, { method: 'GET' });
-  },
-};
-
-async function requestAsync(path: string, options: RequestInit): Promise<any> {
-  try {
-    const response = await fetch(`${getExpoApiBaseUrl()}/v2/${path}`, {
-      ...options,
-      headers: {
-        'Content-Type': 'application/json',
-        ...getAuthHeaders(),
-      },
-    });
-    return await response.json();
-  } catch (err) {
-    await handleApiErrorAsync(err);
+  public async putAsync(path: string, options: RequestOptions): Promise<any> {
+    return await this.requestAsync(path, { method: 'PUT', body: JSON.stringify(options.body) });
   }
-}
 
-async function handleApiErrorAsync(err: any): Promise<void> {
-  if (err instanceof RequestError) {
-    let result: { [key: string]: any };
+  public async postAsync(path: string, options: RequestOptions): Promise<any> {
+    return await this.requestAsync(path, { method: 'POST', body: JSON.stringify(options.body) });
+  }
+
+  public async deleteAsync(path: string): Promise<any> {
+    return await this.requestAsync(path, { method: 'DELETE' });
+  }
+
+  public async getAsync(path: string): Promise<any> {
+    return await this.requestAsync(path, { method: 'GET' });
+  }
+
+  private getAuthHeaders(): Record<string, string> {
+    const token = this.authInfo.accessToken;
+    if (token) {
+      return { authorization: `Bearer ${token}` };
+    }
+    const sessionSecret = this.authInfo.sessionSecret;
+    if (sessionSecret) {
+      return { 'expo-session': sessionSecret };
+    }
+    return {};
+  }
+
+  private async requestAsync(path: string, options: RequestInit): Promise<any> {
     try {
-      result = (await err.response.json()) as { [key: string]: any };
-    } catch {
-      throw new Error(`Malformed api response: ${await err.response.text()}`);
+      const response = await fetch(`${getExpoApiBaseUrl()}/v2/${path}`, {
+        ...options,
+        headers: {
+          'Content-Type': 'application/json',
+          ...this.getAuthHeaders(),
+        },
+      });
+      return await response.json();
+    } catch (err) {
+      await ApiV2Client.handleApiErrorAsync(err);
     }
-    if (result.errors?.length) {
-      throw new ApiV2Error(result.errors[0]);
-    }
-  } else {
-    throw err;
   }
-}
 
-function getAuthHeaders(): Record<string, string> {
-  const token = getAccessToken();
-  if (token) {
-    return { authorization: `Bearer ${token}` };
+  private static async handleApiErrorAsync(err: any): Promise<void> {
+    if (err instanceof RequestError) {
+      let result: { [key: string]: any };
+      try {
+        result = (await err.response.json()) as { [key: string]: any };
+      } catch {
+        throw new Error(`Malformed api response: ${await err.response.text()}`);
+      }
+      if (result.errors?.length) {
+        throw new ApiV2Error(result.errors[0]);
+      }
+    } else {
+      throw err;
+    }
   }
-  const sessionSecret = getSessionSecret();
-  if (sessionSecret) {
-    return { 'expo-session': sessionSecret };
-  }
-  return {};
 }
 
 export function getExpoApiBaseUrl(): string {
