+++ packages/eas-cli/src/commandUtils/context/LoggedInContextField.ts
@@ -1,17 +1,20 @@
-import { legacyGraphqlClient } from '../../graphql/client';
 import { Actor } from '../../user/User';
 import ContextField, { ContextOptions } from './ContextField';
-import { ExpoGraphqlClient } from './contextUtils/createGraphqlClient';
-import { ensureLoggedInAsync } from './contextUtils/ensureLoggedInAsync';
+import { ExpoGraphqlClient, createGraphqlClient } from './contextUtils/createGraphqlClient';
 
 export default class LoggedInContextField extends ContextField<{
   actor: Actor;
   graphqlClient: ExpoGraphqlClient;
 }> {
-  async getValueAsync({
-    nonInteractive,
-  }: ContextOptions): Promise<{ actor: Actor; graphqlClient: ExpoGraphqlClient }> {
-    const actor = await ensureLoggedInAsync({ nonInteractive });
-    return { actor, graphqlClient: legacyGraphqlClient };
+  async getValueAsync({ nonInteractive, sessionManager }: ContextOptions): Promise<{
+    actor: Actor;
+    graphqlClient: ExpoGraphqlClient;
+  }> {
+    const { actor, authenticationInfo } = await sessionManager.ensureLoggedInAsync({
+      nonInteractive,
+    });
+
+    const graphqlClient = createGraphqlClient(authenticationInfo);
+    return { actor, graphqlClient };
   }
 }
