+++ packages/eas-cli/src/commandUtils/context/SessionManagementContextField.ts
@@ -1,0 +1,8 @@
+import SessionManager from '../../user/SessionManager';
+import ContextField, { ContextOptions } from './ContextField';
+
+export default class SessionManagementContextField extends ContextField<SessionManager> {
+  async getValueAsync({ sessionManager }: ContextOptions): Promise<SessionManager> {
+    return sessionManager;
+  }
+}
