+++ packages/eas-cli/src/commands/account/login.ts
@@ -1,13 +1,17 @@
 import EasCommand from '../../commandUtils/EasCommand';
-import { showLoginPromptAsync } from '../../commandUtils/context/contextUtils/ensureLoggedInAsync';
 import Log from '../../log';
 
 export default class AccountLogin extends EasCommand {
   static override description = 'log in with your Expo account';
   static override aliases = ['login'];
 
+  static override contextDefinition = {
+    ...this.ContextOptions.SessionManagment,
+  };
+
   async runAsync(): Promise<void> {
-    await showLoginPromptAsync();
+    const { sessionManager } = await this.getContextAsync(AccountLogin, { nonInteractive: false });
+    await sessionManager.showLoginPromptAsync();
     Log.log('Logged in');
   }
 }
