+++ packages/eas-cli/src/commands/account/logout.ts
@@ -1,13 +1,17 @@
 import EasCommand from '../../commandUtils/EasCommand';
 import Log from '../../log';
-import { logoutAsync } from '../../user/User';
 
 export default class AccountLogout extends EasCommand {
   static override description = 'log out';
   static override aliases = ['logout'];
 
+  static override contextDefinition = {
+    ...this.ContextOptions.SessionManagment,
+  };
+
   async runAsync(): Promise<void> {
-    await logoutAsync();
+    const { sessionManager } = await this.getContextAsync(AccountLogout, { nonInteractive: false });
+    await sessionManager.logoutAsync();
     Log.log('Logged out');
   }
 }
