+++ packages/eas-cli/src/project/__tests__/publish-test.ts
@@ -350,6 +350,7 @@ describe(filterOutAssetsThatAlreadyExistAsync, () => {
         },
       ];
     });
+
     expect(
       (await filterOutAssetsThatAlreadyExistAsync(graphqlClient, [{ storageKey: 'blah' } as any]))
         .length
