+++ packages/eas-cli/src/submit/BaseSubmitter.ts
@@ -2,7 +2,6 @@ import { Platform } from '@expo/eas-build-job';
 
 import { withAnalyticsAsync } from '../analytics/common';
 import { Event, SubmissionEvent } from '../analytics/events';
-import { ExpoGraphqlClient } from '../commandUtils/context/contextUtils/createGraphqlClient';
 import {
   AndroidSubmissionConfigInput,
   IosSubmissionConfigInput,
@@ -79,10 +78,7 @@ export default abstract class BaseSubmitter<
     const platformDisplayName = appPlatformDisplayNames[toAppPlatform(this.ctx.platform)];
     const scheduleSpinner = ora(`Scheduling ${platformDisplayName} submission`).start();
     try {
-      const submission = this.createPlatformSubmissionAsync(
-        this.ctx.graphqlClient,
-        submissionInput
-      );
+      const submission = this.createPlatformSubmissionAsync(submissionInput);
       scheduleSpinner.succeed(`Scheduled ${platformDisplayName} submission`);
       return submission;
     } catch (err) {
@@ -106,7 +102,6 @@ export default abstract class BaseSubmitter<
   }
 
   protected abstract createPlatformSubmissionAsync(
-    graphqlClient: ExpoGraphqlClient,
     input: SubmissionInput<P>
   ): Promise<SubmissionFragment>;
 }
