+++ packages/eas-cli/src/user/__tests__/User-test.ts
@@ -1,43 +1,5 @@
-import fs from 'fs-extra';
-import { vol } from 'memfs';
-
 import { Role } from '../../graphql/generated';
-import { getStateJsonPath } from '../../utils/paths';
-import {
-  Actor,
-  getActorDisplayName,
-  getSessionSecret,
-  getUserAsync,
-  loginAsync,
-  logoutAsync,
-} from '../User';
-
-jest.mock('fs');
-jest.mock('../../api', () => ({
-  api: {
-    postAsync: jest.fn(async () => ({ data: { sessionSecret: 'SESSION_SECRET' } })),
-  },
-  ApiV2Error: jest.requireActual('../../api').ApiV2Error,
-}));
-jest.mock('../../graphql/client', () => ({
-  legacyGraphqlClient: {
-    query: () => {
-      return {
-        toPromise: () =>
-          Promise.resolve({ data: { viewer: { id: 'USER_ID', username: 'USERNAME' } } }),
-      };
-    },
-  },
-}));
-jest.mock('../../graphql/queries/UserQuery', () => ({
-  UserQuery: {
-    currentUserAsync: async () => ({ __typename: 'User', username: 'USERNAME', id: 'USER_ID' }),
-  },
-}));
-
-beforeEach(() => {
-  vol.reset();
-});
+import { Actor, getActorDisplayName } from '../User';
 
 const userStub: Actor = {
   __typename: 'User',
@@ -65,50 +27,6 @@ const robotStub: Actor = {
   accounts: [],
   isExpoAdmin: false,
 };
-
-describe('getUserAsync', () => {
-  it('skips fetching user without access token or session secret', async () => {
-    expect(await getUserAsync()).toBeUndefined();
-  });
-
-  it('fetches user when access token is defined', async () => {
-    process.env.EXPO_TOKEN = 'accesstoken';
-    expect(await getUserAsync()).toMatchObject({ __typename: 'User' });
-  });
-
-  it('fetches user when session secret is defined', async () => {
-    await loginAsync({ username: 'USERNAME', password: 'PASSWORD' });
-    expect(await getUserAsync()).toMatchObject({ __typename: 'User' });
-  });
-});
-
-describe('loginAsync', () => {
-  it('saves user data to ~/.expo/state.json', async () => {
-    await loginAsync({ username: 'USERNAME', password: 'PASSWORD' });
-
-    expect(await fs.readFile(getStateJsonPath(), 'utf8')).toMatchInlineSnapshot(`
-      "{
-        \\"auth\\": {
-          \\"sessionSecret\\": \\"SESSION_SECRET\\",
-          \\"userId\\": \\"USER_ID\\",
-          \\"username\\": \\"USERNAME\\",
-          \\"currentConnection\\": \\"Username-Password-Authentication\\"
-        }
-      }
-      "
-    `);
-  });
-});
-
-describe('logoutAsync', () => {
-  it('removes the session secret', async () => {
-    await loginAsync({ username: 'USERNAME', password: 'PASSWORD' });
-    expect(getSessionSecret()).toBe('SESSION_SECRET');
-
-    await logoutAsync();
-    expect(getSessionSecret()).toBe(null);
-  });
-});
 
 describe('getActorDisplayName', () => {
   it('returns anonymous for unauthenticated users', () => {
